rename: Home 부분 파일 하위에 index.tsx를 두는 방식으로 파일 수정
