+++ src/components/Admin/Home/Todo/index.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { TodoList } from '../../../utils/admin/ArrayItem';
+import { TodoList } from '../../../../utils/admin/ArrayItem';
 
 const Todo = () => {
   return (
